requirements.txt
# Core dependencies
pandas==2.2.3
tweepy==4.14.0
httpx==0.27.0
supabase==2.3.1
python-dotenv==1.0.1
requests==2.32.3
